update xor testing
